// frontend/src/shared/api/schema.d.ts
export interface components {
    schemas: {
// ...
        VersionDetailResponse: {
            /** Version */
            version: number;
            view: components["schemas"]["VersionViewData"];
        };
// ...
        VersionsListResponse: {
            /** Versions */
            versions: components["schemas"]["VersionSummary"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}